Build Nix flake in GitHub Actions (#7)
Changed settings of the rust workflow to check if the program can be built on any branch and pull request.
Added a job build-nix to the rust workflow that checks if the program can be built from Nix flake.
If the job build-nix fails due to an outdated cargoHash, it calculates the correct value and displays it in an error message for easy copying to flake.nix.